refacotr : 인스타그램 연동 기능 계층화
인스타그램 연동기능을 계층분리 하였습니다.
