+++ src/main/java/com/kakaotech/team14backend/oauth2/application/ConnectSuccessHandler.java
@@ -1,0 +1,40 @@
+package com.kakaotech.team14backend.oauth2.application;
+
+import com.kakaotech.team14backend.common.ApiResponse;
+import com.kakaotech.team14backend.common.ApiResponseGenerator;
+import com.kakaotech.team14backend.inner.member.model.Member;
+import com.kakaotech.team14backend.inner.member.repository.MemberRepository;
+import com.kakaotech.team14backend.jwt.dto.TokenDTO;
+import com.kakaotech.team14backend.jwt.service.TokenService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Component;
+
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletResponse;
+
+@Component
+@RequiredArgsConstructor
+public class ConnectSuccessHandler {
+  @Value("${jwt.token-validity-in-seconds-refreshToken}")
+  private int cookieAge;
+
+  private final MemberRepository memberRepository;
+  private final TokenService tokenService;
+  public void execute(HttpServletResponse response, Member member) {
+    TokenDTO tokenDTO = tokenService.createOrUpdateToken(member);
+    response.setContentType("application/json");
+    // 토큰을 HTTP 헤더에 추가
+    response.addHeader("Authorization", tokenDTO.getAccessToken());
+    // 리프레시토큰을 쿠키에 저장한다.
+    String refreshTokenCookieName = "RefreshToken";
+    String refreshTokenCookieValue = tokenDTO.getRefreshToken();
+    Cookie refreshTokenCookie = new Cookie(refreshTokenCookieName, refreshTokenCookieValue);
+    refreshTokenCookie.setHttpOnly(true);  //httponly 옵션 설정
+    refreshTokenCookie.setSecure(true); //https 옵션 설정
+    refreshTokenCookie.setPath("/"); // 모든 곳에서 쿠키열람이 가능하도록 설정
+    refreshTokenCookie.setMaxAge(cookieAge); //쿠키 만료시간 설정
+    response.addCookie(refreshTokenCookie);
+  }
+}
